supply-block-chains-master/requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pycparser==2.18
pycryptodomex==3.6.4
pyOpenSSL==18.0.0
PySocks==1.6.8
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.10
urllib3==1.23
Werkzeug==0.14.1
WTForms==2.2.1
